src/selenium/requirements.txt
selenium
beautifulsoup4
requests
gitpython

## Run `pip install selenium beautifulsoup4 requests gitpython`
